flask/requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
decorator==4.4.2
h11==0.14.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy
opencv-python==4.8.1.78
outcome==1.2.0
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.6
webdriver-manager==4.0.1
wsproto==1.2.0
